tuples

+++ python/import/calcmain.py
@@ -6,5 +6,5 @@ def sub(a,b):
 def mul(a,b):
     return a*b
 def div(a, b):
-    return a/b
+    return int(a/b)
